requirements.txt
requests
pygithub
git
